requirements.txt
pinecone-client
openai
gradio
python-dotenv
